react-drf-blog-api/pyproject.toml
[tool.poetry]
name = "drf-blog"
version = "1.0.0"
description = "DRF Blog"
license = "MIT"
authors = ["Jake <renownkim@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/kimfame/react-drf-blog.git"
repository = "https://github.com/kimfame/react-drf-blog.git"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Topic :: Internet :: WWW/HTTP",
]
packages = [{include = "drf_blog"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
django-extensions = "^3.2.3"
django-ckeditor = "^6.5.1"
django-cors-headers = "^4.1.0"


[tool.poetry.group.deploy.dependencies]
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"


[tool.poetry.group.test.dependencies]
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
